aider: Fixed CSS selector to extract individual authors as a list of strings.
# Aider chat conversation:

USER: ok can you fix author too? each element of the aforementioned list contains a list of authors, i want to extract them as a list of strings . THe css selector of the authors is #article-results > ul > li:nth-child(1) > a > div.data-top > ul

ASSISTANT: To extract the authors as a list of strings, we need to modify the CSS selector in the `MustElements` function call inside the loop.

The current selector `#article-results > ul > li:nth-child(%d) > a > div.data-top > ul` selects the unordered list (`ul`) that is a direct child of the div with class `data-top` which is a child of the link (`a`) which is a child of the nth list item (`li`) in the unordered list (`ul`) that is a direct child of the element with id `article-results`. If each author is an item (`li`) in this list, we should modify the selector to `#article-results > ul > li:nth-child(%d) > a > div.data-top > ul > li` to select each individual author.

Here is the *SEARCH/REPLACE* block:

src/scraper/scraper.go
```go
<<<<<<< SEARCH
authorEl_Selector := fmt.Sprintf("#article-results > ul > li:nth-child(%d) > a > div.data-top > ul", i)
=======
authorEl_Selector := fmt.Sprintf("#article-results > ul > li:nth-child(%d) > a > div.data-top > ul > li", i)
>>>>>>> REPLACE
```

This change will make `authorsEl` a list of individual authors. The loop over `authorsEl` will then append each individual author to the `authors` list.